backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
rank-bm25==0.2.2
sentence-transformers==3.1.1
faiss-cpu==1.8.0
pythainlp==5.0.4
numpy==1.26.4
requests==2.32.3
